Refactored all classes as services.
