+++ RadioSharp.App/Data/IDatabaseService.cs
@@ -1,0 +1,11 @@
+﻿using RadioSharp.App.Models;
+
+namespace RadioSharp.App.Data
+{
+    public interface IDatabaseService
+    {
+        void AddRadio(RadioStation radio);
+        IList<RadioStation> GetRadios();
+        void InitDatabase();
+    }
+}
