+++ RadioSharp.App/Menus/IMenuService.cs
@@ -1,0 +1,11 @@
+﻿
+namespace RadioSharp.App.Menus
+{
+    public interface IMenuService
+    {
+        Task DisplayPlayBackMenuAsync(bool lastPlayed = false);
+        void DisplayRadioMenu(int page);
+        Task DisplaySearchMenuAsync();
+        void DrawAppLogo();
+    }
+}
