+++ RadioSharp.App/Player/IRadioPlayer.cs
@@ -1,0 +1,9 @@
+﻿using RadioSharp.App.Models;
+
+namespace RadioSharp.App.Player
+{
+    public interface IRadioPlayer
+    {
+        void PlayStream(RadioStation selectedRadio, int choice);
+    }
+}
